+++ newton/_src/solvers/vbd_palatial/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 The Newton Developers
+# SPDX-License-Identifier: Apache-2.0
+
+"""Compatibility exports for the Palatial VBD solver name."""
+
+from .solver_vbd import SolverVBD as SolverVBDPalatial
+
+__all__ = [
+    "SolverVBDPalatial",
+]